move commands ts files
