+++ packages/cli/src/Doctor.ts
