+++ packages/cli/src/Validate.ts
